+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  * [GH-329](https://github.com/rajanadar/VaultSharp/issues/329) kv2/metadata: `ReadSecretPathsAsync` to use HTTP `GET` method and `?list=true` instead of non-standard HTTP verb `LIST`
  * [GH-334](https://github.com/rajanadar/VaultSharp/issues/334) kv2/metadata: `ReadSecretPathsAsync` allows empty path value to list all secrets on the `mountPoint`
  * [GH-342](https://github.com/rajanadar/VaultSharp/issues/342) /sys/policies/password/:name/generate password generation api.
+ * [GH-317](https://github.com/rajanadar/VaultSharp/issues/317) Read Default Issuer Certificate Chain API.
 
 ## 1.13.0.1 (April 03, 2023)